+++ CHANGELOG.md
@@ -13,6 +13,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.0.1] - 2018-09-21
 ### Added
+- add editor schematic
+- add idea schematic
+- add bin schematic
 - add basic package.json
 - add yarn add as add schematic
 - add yarn add --dev as addDev schematic